fix(ws): update component structure to fix create ws button rendering (#364)

fix tests


fix tests
